requirements.txt
django-ninja==1.3.0
uvicorn==0.30.6
django-allow-cidr==0.7.1
django-environ==0.11.2
orjson==3.10.7
psycopg==3.2.3
httpx
boto3==1.37.28
django-storages==1.14.1
django==4.2.20
PyJWT==2.10.1